Correct using of memset
struct SkipNode* update[SL_MAX_LEVEL + 1]
It is array of pointers. Every pointer takes 4 or 8 bytes. It depends from machine.
The third argument of function memset is a number of bytes which should be set to the second argument.
That is why this line
memset(update, 0, (SL_MAX_LEVEL + 1) * sizeof(SkipNode*));
is correct more than
memset(update, 0, SL_MAX_LEVEL + 1);